requirements.txt
pandas
pyarrow
tqdm
pandas-ta
ccxt
python-telegram-bot
schedule